requirements.txt
ccxt>=4.3.89
yfinance>=0.2.40
pandas>=2.2.2
numpy>=2.0.0
ta>=0.11.0
websockets>=12.0
fastapi>=0.111.0
uvicorn>=0.30.1
python-dotenv>=1.0.1
httpx>=0.27.0
aiohttp>=3.9.5
pydantic>=2.7.1
rich>=13.7.1
plotly>=5.22.0
python-dateutil>=2.9.0
apscheduler>=3.10.4
